requirements.txt
dnspython==2.7.0
pymongo==4.13.2
python-dotenv==1.1.1
resend==0.8.0
requests==2.31.0
